refactor: home 의존성 제거 및 expense-summary 기반으로 구조 개선
