+++ apps/web/src/entities/date/index.ts
@@ -1,0 +1,1 @@
+export { useDateStore } from './model/useDateStore';
