+++ apps/web/src/features/expense-summary/index.ts
@@ -1,0 +1,3 @@
+export { useExpenseSummaryData } from './model/useExpenseSummaryData';
+export { MonthlyExpenseInfo } from './ui/MonthlyExpenseInfo';
+export type { MonthlyExpenseInfoProps } from './ui/MonthlyExpenseInfo';
