+++ apps/web/src/features/expense-summary/ui/MonthlyExpenseInfo/MonthlyExpenseInfo.css.ts
