--- apps/web/src/features/homeExpenseData/index.ts
@@ -1,1 +1,0 @@
-export { useHomeExpenseData } from './model/useHomeExpenseData';
