+++ apps/web/src/widgets/home/ui/ExpenseSummary/ExpenseSummary.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Text } from '@/shared/ui/text';
-import { getLoadingClass } from '../../lib/getLoadingClass';
+import { getLoadingClass } from '@/shared/lib/getLoadingClass';
 import * as styles from './ExpenseSummary.css';
 
 export interface ExpenseSummaryProps {
